+++ anvay/anvay/src/main/java/com/cts/mfrp/anvay/service/ApplicationService.java
@@ -6,8 +6,8 @@ import com.cts.mfrp.anvay.entity.LeadershipApplication;
 
 public interface ApplicationService {
     LeadershipApplication createApplication(LeadershipApplication application);
-    LeadershipApplication getApplicationById(Long applicationId);
-    List<LeadershipApplication> getApplicationsByClubId(Long clubId);
-    LeadershipApplication updateApplication(Long applicationId, LeadershipApplication application);
-    void deleteApplication(Long applicationId);
+    LeadershipApplication getApplicationById(Integer applicationId);
+    List<LeadershipApplication> getApplicationsByClubId(Integer clubId);
+    LeadershipApplication updateApplication(Integer applicationId, LeadershipApplication application);
+    void deleteApplication(Integer applicationId);
 }
